% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{check_addin_deps}
\alias{check_addin_deps}
\title{Check if needed packages are installed}
\usage{
check_addin_deps()
}
\value{
a boolean
}
\description{
Check if needed packages are installed
}
\keyword{internal}
