% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect outliers}
\usage{
detect_outliers(
  .data,
  test_type = c("incremental", "kinetics"),
  vo2_column = "VO2",
  cleaning_level = 0.95,
  cleaning_baseline_fit,
  protocol_n_transitions,
  protocol_baseline_length,
  protocol_transition_length,
  method_incremental = c("linear", "anomaly"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{.data}{Data retrieved from \code{read_data()} for a \strong{kinetics} test, or
the data retrieved from \code{incremental_normalize()} for a \strong{incremental} test.}

\item{test_type}{The test to be analyzed. Either 'incremental' or 'kinetics'.}

\item{vo2_column}{The name (quoted) of the column containing the absolute oxygen uptake (VO2) data. Default to \code{VO2}.}

\item{cleaning_level}{A numeric scalar between 0 and 1 giving the confidence level for the intervals to be calculated. Default to \code{0.95}.}

\item{cleaning_baseline_fit}{For \strong{kinetics} test only. A vector of the same length as the number in \code{protocol_n_transitions}, indicating what kind of fit to perform for each baseline. Vector accepts characters either 'linear' or 'exponential'.}

\item{protocol_n_transitions}{For \strong{kinetics} test only. Number of transitions performed.}

\item{protocol_baseline_length}{For \strong{kinetics} test only. The length of the baseline (in seconds).}

\item{protocol_transition_length}{For \strong{kinetics} test only. The length of the transition (in seconds).}

\item{method_incremental}{The method to be used in detecting outliers from the
incremental test. Either 'linear' or 'anomaly'. See \code{Details}.}

\item{verbose}{A boolean indicating whether messages should be printed in the console. Default to \code{TRUE}.}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
It detects outliers based on prediction bands for the given level of confidence provided.
}
\details{
TODO
}
