% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlstools.R
\name{get_residuals}
\alias{get_residuals}
\title{Get residuals}
\usage{
get_residuals(.model)
}
\arguments{
\item{.model}{A model of class \code{nls}.}
}
\value{
a \link[tibble:tibble-package]{tibble} containing the data passed to augment, and additional columns:
\item{.fitted}{The predicted response for that observation.}
\item{.resid}{The residual for a particular point.}
\item{standardized_residuals}{Standardized residuals.}
\item{sqrt_abs_standardized_residuals}{The sqrt of absolute value of standardized residuals.}
\item{lag_residuals}{The lag of the \code{.resid} column for plotting auto-correlation.}
}
\description{
Computes residuals from the VO2 kinetics model.
}
