% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrt.R
\name{incremental_mrt}
\alias{incremental_mrt}
\title{Mean Response Time (MRT)}
\usage{
incremental_mrt(
  .data,
  vo2_column = "VO2",
  method = c("linear", "caen", "iannetta"),
  GET,
  slope_2_cwr,
  target_work_rate
)
}
\arguments{
\item{.data}{The data retrieved from \code{incremental_normalize()}.}

\item{vo2_column}{The name (quoted) of the column containing the absolute oxygen uptake (VO2) data. Default to \code{VO2}.}

\item{method}{The method to be used for MRT estimation. See Details.}

\item{GET}{The VO2 associated with the gas exchange threshold. It must be in the same units as your raw data (either mL/min or L/min).}

\item{slope_2_cwr}{TODO}

\item{target_work_rate}{TODO}
}
\value{
a \link[tibble:tibble-package]{tibble} containing one row and the following columns for the linear method:
\item{data}{The raw data containing and additional column with the data from the double-linear fit.}
\item{MRT_time}{The MRT estimation, in seconds.}
\item{MRT_work_rate}{The MRT estimation, in watts}
\item{plot}{A \code{ggplot2} object with the plot showing the MRT.}
}
\description{
Calculates the mean response time to be used for correction of the work rate estimation at any given exercise intensity.
See Details.
}
\details{
TODO
\subsection{mention that these are only for intensities < RCP}{
}
}
\examples{
## get file path from example data
path_example <- system.file("ramp_cosmed.xlsx", package = "whippr")

## read data
df <- read_data(path = path_example, metabolic_cart = "cosmed")

## normalize incremental test data
ramp_normalized <- df \%>\%
 incremental_normalize(
   .data = .,
   incremental_type = "ramp",
   has_baseline = TRUE,
   baseline_length = 240,
   work_rate_magic = TRUE,
   baseline_intensity = 20,
   ramp_increase = 25
 )

## calculate the MRT
incremental_mrt(.data = ramp_normalized, vo2_column = "VO2", method = "linear", GET = 1800)
}
