% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlstools.R
\name{model_diagnostics}
\alias{model_diagnostics}
\title{Model diagnostics}
\usage{
model_diagnostics(.residuals_tbl)
}
\arguments{
\item{.residuals_tbl}{The data retrived from \code{get_residuals()}.}
}
\value{
a patchwork object
}
\description{
Plots different model diagnostics for checking the model performance.
}
