% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{new_whippr_tibble}
\alias{new_whippr_tibble}
\title{Construct a new tibble with metadata}
\usage{
new_whippr_tibble(.data, metadata)
}
\arguments{
\item{.data}{A data frame}

\item{metadata}{Metadata to be passed along with the data}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Construct a new tibble with metadata
}
\keyword{internal}
