% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_first_breath}
\alias{normalize_first_breath}
\title{Normalize first breath}
\usage{
normalize_first_breath(.data)
}
\arguments{
\item{.data}{Breath-by-breath data.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
This is needed specially when the data gets filtered. For example, if the data file does not only contain
the baseline and transitions performed, we will have to normalize the time column.
This function will make sure that in case the first breath does not start at zero, it will create a zero data point,
duplicating the first breath. This will make sure the data does not get shifted (misalignment).
}
