% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-outliers.R
\name{outliers_anomaly}
\alias{outliers_anomaly}
\title{Anomaly method for detecting outliers from an incremental test}
\usage{
outliers_anomaly(.data, time_column, vo2_column, cleaning_level)
}
\arguments{
\item{.data}{The data retrieved from \code{incremental_normalize()}.}

\item{time_column}{The name (quoted) of the column containing the time. Depending on the language of your system, this column might not be "t". Therefore, you may specify it here.}

\item{vo2_column}{The name (quoted) of the column containing the absolute oxygen uptake (VO2) data.}

\item{cleaning_level}{A numeric scalar between 0 and 1 giving the confidence level for the intervals to be calculated.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Function for internal use only. It will not be exported.
}
\keyword{internal}
