% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averages.R
\name{perform_average}
\alias{perform_average}
\title{Perform average on second-by-second data}
\usage{
perform_average(
  .data,
  time_column = "t",
  type = c("bin", "rolling"),
  bins = 30,
  rolling_window = 30
)
}
\arguments{
\item{.data}{The second-by-second data retrieved from \code{interpolate()}.}

\item{time_column}{The name (quoted) of the column containing the time. Default to "t".}

\item{type}{The type of the average to perform. Either \code{bin} or \code{rolling}.}

\item{bins}{If bin-average is chosen, here you can specify the size of the bin-average, in seconds. Default to 30-s bin-average.}

\item{rolling_window}{If rolling-average is chosen, here you can specify the rolling-average window, in seconds. Default to 30-s rolling-average.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
This function performs either a bin- or a rolling-average on the interpolated data.
You must specify the \code{type} of the average before continuing.
}
\examples{
\dontrun{
## read data
df <- read_data(path = "~/Documents/data.xlsx", metabolic_cart = "cosmed")

## interpolate and perform 30-s bin-average
df \%>\%
 interpolate() \%>\%
 perform_average(type = "bin", bins = 30)

## interpolate and perform 30-s rolling-average
df \%>\%
 interpolate() \%>\%
 perform_average(type = "rolling", rolling_window = 30)
}
}
