% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incremental.R
\name{plot_incremental}
\alias{plot_incremental}
\title{Plot incremental test work rate}
\usage{
plot_incremental(.data)
}
\arguments{
\item{.data}{data retrieved from \code{incremental_normalize()}.}
}
\value{
a ggplot object
}
\description{
Visualize what was done during the process of deriving the work rate from the incremental test protocol
}
