% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{plot_outliers}
\alias{plot_outliers}
\title{Plot outliers}
\usage{
plot_outliers(.data)
}
\arguments{
\item{.data}{The data retrieved from \code{detect_outliers()}.}
}
\value{
a patchwork object
}
\description{
Plot outliers
}
