% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{print.whippr}
\alias{print.whippr}
\title{Whippr print method}
\usage{
\method{print}{whippr}(x, ...)
}
\arguments{
\item{x}{A tibble with class 'whippr'}

\item{...}{Extra arguments, not used.}
}
\description{
Whippr print method
}
