% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\name{process_data}
\alias{process_data}
\title{Process data for VO2 kinetics fitting}
\usage{
process_data(.data_outliers, protocol_baseline_length, fit_bin_average)
}
\arguments{
\item{.data_outliers}{The data retrived from \code{detect_outliers()}.}

\item{protocol_baseline_length}{The length of the baseline (in seconds).}

\item{fit_bin_average}{The bin average to be performed for the final fit.}
}
\value{
a \link[tibble:tibble-package]{tibble} with the time-aligned, ensembled-averaged, and bin-averaged data.
}
\description{
It removes the outliers detected through \code{detect_outliers()}, interpolates each transition,
ensemble-averages all the transitions into one, performs a bin-average, and normalizes the time column
(time zero will indicate the end of baseline and the start of the transition phase).
}
\details{
TODO
}
