% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_data}
\alias{read_data}
\title{Read data from metabolic cart}
\usage{
read_data(path, metabolic_cart = c("cosmed", "cortex"))
}
\arguments{
\item{path}{Path to read the file from.}

\item{metabolic_cart}{Metabolic cart that was used for data collection. Currently, only 'cosmed' and 'cortex' are supported.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
It reads the raw data exported from the metabolic cart.
}
