% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{run_manual_cleaner}
\alias{run_manual_cleaner}
\title{Manual data cleaner}
\usage{
run_manual_cleaner(.data, width = 1200, height = 900)
}
\arguments{
\item{.data}{The data to be manually cleaned. The first column will be always treated as the x-axis.}

\item{width}{The width, in pixels, of the window.}

\item{height}{the height, in pixels, of the window.}
}
\value{
The code to reproduce the manual data cleaning.
}
\description{
Usually manual data cleaning should be avoided. However, sometimes in gas exchange data
there is the need to delete a few clear "bad breaths" (noise). In these situations you may use this function.
Although it is encouraged that you use the \code{detect_outliers()} function, you may use this function at your own risk.
This function can also be used to clean other kind of data, like heart rate data.
}
