% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-theme.R
\name{theme_whippr}
\alias{theme_whippr}
\title{Whippr ggplot2 theme}
\usage{
theme_whippr()
}
\value{
a ggplot2 object
}
\description{
This theme was inspired by the plots from the Acta Physiologica Journal
}
