% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whippr-package.R
\docType{package}
\name{whippr-package}
\alias{whippr}
\alias{whippr-package}
\title{whippr: Tools for Manipulating Gas Exchange Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

'whippr' provides a set of tools for manipulating gas exchange data from cardiopulmonary exercise testing.
}
\author{
\strong{Maintainer}: Felipe Mattioni Maturana \email{felipe.mattioni@med.uni-tuebingen.de} (\href{https://orcid.org/0000-0002-4221-6104}{ORCID})

}
\keyword{internal}
