% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-incremental.R
\name{work_rate_ramp}
\alias{work_rate_ramp}
\title{Work rate for a ramp-incremental test}
\usage{
work_rate_ramp(.data, baseline_intensity, ramp_increase)
}
\arguments{
\item{.data}{The data with recognized protocol phases}

\item{baseline_intensity}{The baseline intensity}

\item{ramp_increase}{The ramp increase, in watts per minute}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
This function produces the work rate throughout a ramp-incremental test given the procotol
}
\keyword{internal}
