% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate from breath-by-breath into second-by-second}
\usage{
interpolate(.data, time_column = "t")
}
\arguments{
\item{.data}{Data retrieved from \code{read_data()}.}

\item{time_column}{The name (quoted) of the column containing the time. Default to "t".}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
This function interpolates the data based on the time column. It takes the breath-by-breath data
and transforms it into second-by-second.
}
\examples{
\dontrun{
## read data
df <- read_data(path = "~/Documents/data.xlsx", metabolic_cart = "cosmed")

df_interpolated <- df \%>\%
   interpolate()
}
}
