% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_data}
\alias{read_data}
\title{Read data from metabolic cart}
\usage{
read_data(
  path,
  metabolic_cart = c("cosmed", "cortex", "nspire", "parvo"),
  time_column = "t"
)
}
\arguments{
\item{path}{Path to read the file from.}

\item{metabolic_cart}{Metabolic cart that was used for data collection. Currently, 'cosmed', 'cortex', 'nspire', and 'parvo' are supported.}

\item{time_column}{The name (quoted) of the column containing the time. Depending on the language of your system, this column might not be "t". Therefore, you may specify it here.  Default to "t".}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
It reads the raw data exported from the metabolic cart.
}
