% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whippr-package.R
\docType{package}
\name{whippr-package}
\alias{whippr}
\alias{whippr-package}
\title{whippr: Tools for Manipulating Gas Exchange Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Set of tools for manipulating gas exchange data from cardiopulmonary exercise testing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fmmattioni/whippr}
  \item Report bugs at \url{https://github.com/fmmattioni/whippr/issues}
}

}
\author{
\strong{Maintainer}: Felipe Mattioni Maturana \email{felipe.mattioni@med.uni-tuebingen.de} (\href{https://orcid.org/0000-0002-4221-6104}{ORCID})

}
\keyword{internal}
